requirements.txt
fastapi
uvicorn
pydantic
openai
chromadb
sentence-transformers
python-dotenv